+++ rectipy/__init__.py
@@ -33,4 +33,4 @@ __version__ = "0.2.0"
 
 from .network import Network
 from .utility import random_connectivity, circular_connectivity, input_connections
-from .utility import wta_score
+from .utility import wta_score, readout